All functions working. Refactor branch nexxt

+++ delete_upload.php
@@ -18,13 +18,17 @@
             echo "Fatal error";
             header("Location: department.php?error=fatal");
         }else{
+            $upload = $checkstmt->fetch();
+            $file = "uploads/".$upload['department']."/".$upload['filename'];
+            echo $file;
+            if (!unlink($file)) {
+                header("Location: department.php?error=inaccessible_file");
+            } else {
                 $sql = 'DELETE FROM uploads where id = :id';
                 $stmt = $connection->prepare($sql);
                 $stmt->execute(['id' => $id]);
-                echo "Upload Deleted";
-                header("Location: department.php");
+                header("Location: department.php?success=file_deleted");
+            }
         }
-
-        
     }   
 ?>
